pyproject.toml
[project]
name = "issue-fixer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "anthropic>=0.66.0",
    "dotenv>=0.9.9",
    "git-python>=1.0.3",
    "requests>=2.32.5",
]
